requirements.txt
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.10.4
autopep8==2.3.1
certifi==2024.12.14
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
logging==0.4.9.6
openai==1.58.1
pycodestyle==2.12.1
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
python-telegram-bot==21.9
pytz==2025.1
redis==5.2.1
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzlocal==5.2
